devs/claude/requirements.txt
streamlit               # ==1.26.0
numpy                   #  ==1.24.3
pandas                  #  ==2.0.3
scikit-learn            #  ==1.3.0
plotly                  #  ==5.15.0
sentence-transformers   #  ==2.2.2
umap-learn              #  ==0.5.3
streamlit-aggrid        #  ==0.3.4
